add function for calculating AccurateRip disc IDs
Use the function to print the AccurateRip IDs in cdtoc.py and mbtoc.py.

Add GPLv3 license. The function is adapted from source code of Whipper.
While it's not copied 1:1, I guess it's appropriate that I use the same
license for this project. I intend to copy some more stuff in the future
anyway.

The function seems to work in my testing so far, but it needs to be
tested with CDs containing data tracks (mixed mode or enhanced CDs).