fix calculation of AccurateRip disc IDs
It appears that the TOC should always be adjusted by 150 sectors, not by
the offset of the first track (i.e. the offset of the first track is not
always zero).

The previous version seemed to be working, because the offset of 150
sectors for the first track is very common, but it's not a rule. Some
of my CDs have larger offsets of the first track, and for them I was
getting incorrect results.

Incidentally, these are the same CDs where cdparanoia in batch mode rips
track 0 which is just a fraction of a second of silence.