add accuraterip checksum C module
Copy C code for calculating AccurateRip checksums from Whipper, with
minor adaptations. Add setup.py so that it compiles automatically, and
add Python wrapper for C function so it's actually usable and documented
in Python.

The relative input in the wrapper is necessary, because the module is
not installed system-wide at this stage.

Create AccurateRip submodule which provides ID and checksum calculation.

Ignore files and directories created by setup.py.