add breakdown of AccurateRip binary data
The format is a bit different from what I initially thought, because I
wasn't aware of frame 450 checksums used for offset detection. But after
some experiments and reading forums and code I think I have the right
idea about the data format now.

ODT format is probably not the best for storing this, but I needed to
get this done quickly. If I ever need this example in another format,
I'll change it.