parse full AccurateRip disc data
AccurateRip disc data is composed of one or more responses, each
starting with a 13-byte header followed by a set of 9-byte tracks. So
reading responses one by one should consume the disc data completely:
there will be exactly zero bytes left at the end. The algorithm relies
on this to detect when to stop parsing.

Obviously this will work only when received disc data has the correct
number of bytes. In case of data corruption (missing or extra bytes)
either sanity checking of headers will raise a ValueError, or there will
be an IndexError or struct.error raised when unpacking binary data.