add arver-discinfo
Implement a program for displaying disc info and AccurateRip data of a
compact disc (either for disc in drive, or fetched from MusicBrainz by
disc ID). This obsoletes cdtoc.py and mbtoc.py, which are removed.

arver.py is also removed, because it makes no sense to have a program
which does not work, and it was evolving toward becoming arver-discid
anyway.