add DiscData class
DiscData is a wrapper around a list of Responses. It is convenient to
have such object which can be handled as a whole. Because of that it is
possible to move Disc.format_responses() to __str__() of DiscData, and
to directly print DiscData instance where needed. It also opens the way
to representing AccurateRip data in a way more suitable for verification
of ripped files.

Rename the argument of Disc constructor to disc_info. That way the name
does not clash with class attribute, and it is consistent with naming of
functions which obtain disc information.

Rename Disc.get_responses() to fetch_disc_data(). Adapt discinfo.py to
the changes: ValueError is not thrown anymore when fetching AccurateRip
data fails, None is returned instead.

Update docstrings where needed.