represent DiscData as a dictionary
Add method for converting DiscData object to a dictionary.

The dictionary is track-based, i.e. it is a collection of Tracks, all
of which store checksums collected from one or more Responses. Since
checksum is a dictionary key, it will be easy to check if particular
checksum is assigned to a given track, and with what confidence.

Implement __repr__() in DiscData which returns the data in JSON format
using the dictionary conversion method.