show track listing in discinfo.py
Calculate lengths of tracks in mm:ss.ff format and in sectors, and
display the full track listing in the string representation of Disc,
instead of only printing the number of tracks.

Detect and report possible HTOA by checking if index 1 of the first
track begins immediately after the lead in section. If the first track
begins on a later frame it doesn't necessarily mean that the disc really
contains hidden audio, but the frames before index 1 of the first track
can be ripped as track 0 by cdparanoia. In most cases this will be just
a fraction of second of silence, though.

HTOA cannot be verified with AccurateRip database, so it is displayed
only because the information of its existence is available.