correct usage of AccurateRip checksums
The program has been working under assumption that the version of track
checksum acquired from AccurateRip database is known (i.e. it is clear
whether the checksum is v1 or v2 type). This is not the case, the type
of checksum is not specified in the database. It is not possible to know
the checksum version until checksum of specified type is calculated and
matched to a database entry.

Rename variables holding the checksums and update docstring of Track
class. Ignore frame 450 checksums for now and use the track checksums
correctly. Drop frame 450 checksums and checksum version from dictionary
returned by DiscData.make_dict() method.

Don't print frame 450 checksums when printing Response objects.