create disc.utils module
The module holds two constants and a function for converting track
length in frames to mm:ss.ff format. They will be used in the rip
module. Refactor existing code to use disc.utils.

All of this should be available for import straight from disc module,
but it doesn't work because of naming conflict between disc package and
disc module. I'll figure out how to deal with this correctly some day,
for now this workaround is needed to avoid code duplication.