calculate checkums of audio files
Extend the Rip class so that it extracts all necessary information from
audio files (only WAV supported at the moment), and calculates copy CRCs
and AccurateRip checksums.

Define a type for audio file properties (again, only WAV format). Read
all the required parameters by opening the file just once, instead of
opening several times to get several parameters. Add one more criterion
when checking if a file is sourced from a CD rip: even division by 588
samples.

Do not calculate checksums in WavFile constructor. One thing is that
this calculation is too "heavy". Another problem is that it's not
possible to calculate AccurateRip CRCs without knowing the track number
and the total number of tracks, and this information is only known once
Rip object is fully constructed. Provide setter-type methods to perform
these calculations in WavFile, and implement a method for calculating
all checksums in Rip.

Provide __str__() methods and display the data of all files in a table.
Ensure the table is aligned even when the file names are too long:
shorten the names to the length of the "file name" column.

Finally, reimplement ar.py using Rip class.