handle AccurateRip checksum calculation failure
When accuraterip.compute() fails to calculate AccurateRip checksums
it returns None instead of an int. This can happen when the audio file
cannot be loaded, or is not CDDA (not stereo, with wrong bit depth or
sampling rate). This case was not handled, and resulted in a crash when
AccurateRip checksum was printed.

Change the Python wrapper for compute() so that it returns a pair of
zeros in case of error. The condition checks if any of the returned
checksums is None, and this is sufficient. The calculation will either
succeed for both checksum types, or it will fail for both - there is no
middle ground.

Using zero as a placeholder for an incorrect checksum seems reasonable,
because AccurateRip database uses the same value for missing checksums.
I'm not sure if it's possible to calculate a zero value checksum for a
track with non-zero length anyway.