add arver-dbar-parser
Add program for parsing and printing AccurateRip disc data stored in a
dBAR file cached by EAC or Whipper.

This is hacky. There is no disc in drive or MusicBrainz DiscID to get
the number of tracks and to calculate disc IDs from. AccurateRipParser
constructor requires these parameters, so I'm implementing a child class
with an __init__() which only requires a file path, and overriding the
header validator method so that it doesn't actually validate data read
from the file against number of tracks and disc IDs we don't have.