detect mixed-mode (Yellow Book) CDs
Implement function for checking if disc is mixed-mode or not. Use a
simple test based on the number of sessions and the type of the first
track to determine if the disc is a mixed-mode CD.

Add field for holding information about mixed mode to the DiscInfo
class. This information will be used to correctly process AccurateRip
responses for mixed-mode CDs.