use DiscInfo class in arver-discinfo
Change print_table() method to _format_tracklist() similar to method
previously implemented in arver.disc.disc. Change __str__() of Track
to match previously used format (print fewer columns), and use it to
implement __str__() in DiscInfo.

Adapt arver-discinfo to use the new DiscInfo class. At that point the
executable part of arver.disc.info becomes unnecessary and is removed.
The only change in arver-discinfo visible to the end user is that now
it prints the disc type.