fix treatment of zero checksums in AccurateRip response
Don't treat zero checksum in AccurateRip response as an invalid value.
Only consider confidence value when checking if a checksum should be
used for verification or not.

Checksum of 0x00000000 is actually a valid checksum if the track only
contains digital silence (e.g. silent tracks on "Dark Side of the Spoon"
or "Houses of the Molé"). In theory, it seems possible that AccurateRip
could return a zero checksum with non-zero confidence value. Not sure
how probable that is, but ARver should handle such situation.