handle tracks without checksums in AccurateRip response
AccurateRip doesn't store checksums of all tracks. Checksums are not
kept e.g. when track is short (less than 10 seconds, but I'm not sure
what the exact threshold value is). This was causing false negative
verification because there was no way to know that failure to match file
checksum to AccurateRip checksum happens because there is no AccurateRip
checksum to match to.

Allow track verification result to be non-binary (success / failure / no
data). When verifying checksums first check if AccurateRip response for
a given track actually contains any checksums, and report "no data"
status if it doesn't. Print a message about unknown checksums in the
verification summary.