fix multisession detection with track 1 pregap
The error was caused by first session LSN being greater than the last
session LSN (e.g. 32 > 0). Change LSN comparison to require that for
multisession CDs the last session LSN must be greater than the first
session LSN (not only different from it).

pycdio API provides device.get_last_session() method, but there doesn't
seem to be a method for getting the LSN of the first session. To work
around that limitation the code assumed that the LSN of the first track
is equivalent to the LSN of the first session. This assumption is not
valid for physical CDs with track 1 pregap.

This could have been found only by manual testing with a physical CD
with track 1 pregap. This has also been tested with multisession (i.e.
enhanced) CDs, and it seems to give correct results for all disc types.