add a workaround for possible libcdio bug
get_last_lsn() method doesn't work correctly for track 99. It looks like
a bug in libcdio: get_last_lsn() implementation obtains the first LSN of
the next track, subtracts 1 and returns the resulting number. For track
99 this results in an attempt to get the first LSN of track 100, which
fails (CDDA standard only allows 99 tracks on a CD).

Because track 99 must be the last track on CD, it's immediately followed
by lead out. The LBA offset of lead out is known, and it can be used to
calculate the last LSN of the final track (convert lead out LBA to LSN
and subtract 1 from it).

Tested with "Broken" by Nine Inch Nails. With this change ARver doesn't
crash and calculates correct AccurateRip and MusicBrainz disc IDs.

This is yet another problem that was only possible to discover by manual
testing with a physical medium. The bug did not manifest in v0.4.1 and
earlier versions, because the disc ID implementation is based on discid
library in these versions.