update pycdio version
Use pycdio 0.21 or newer. ARver has been tested with pycdio versions
0.21 and 2.1.1 (latest), and as far as ARver is concerned these versions
are compatible despite the major version change.

The issue with get_last_lsn() of track 99 is seen in versions 0.21 and
2.1.1, so the workaround is still necessary. It appears to be a problem
in libcdio C library, not in pycdio Python bindings.

ARver has been tested with libcdio 0.93 and 2.1.0 (latest). The problem
exists in both versions.