print track results gradually as they are calculated
Change the loop in verify() so that it calls functions for calculating
checksums instead of relying on the checksums being already prepared by
calculate_checksums() call.

The outcome is that arver prints track results as they are calculated,
without a long delay between fetching AccurateRip data and displaying
final verification result. This also fixes several issues reported by
Pylint related to accessing protected members of WavFile.

calculate_checksums() method is still used in arver-ripinfo. The reason
this method is still needed is that one must know the track number and
the total number of tracks to correctly calculate AccurateRip checksums.
It's easy to know when iterating over a list of WavFile objects like
calculate_checksums() does, but wouldn't be possible in a method of
WavFile which doesn't know its own track number or the total number
of tracks.