rename get_disc() and move it to disc.info module
The disc.info module seems to be the appropriate place to keep this
helper function in. This avoids "horizontal" imports between the entry
points.

Change the function so that it expects a single optional string instead
of a list of strings. Update entry points to make argparse return one
string instead of a list. Update function docstring and add type hints.

This fixes a small presentation bug where an invalid disc ID was printed
as a one element list in the error message.