adjust image URL for PyPI
Use a direct link to the raw image at GitHub at a given tag.

The upside is that the animated GIF will be displayed at PyPI. The
downside is that the version string must now be updated in two places,
that the image will be stuck at specified ref between updates, and
that the image will not show either at GitHub nor at PyPI until a tag
is pushed (this would not be a problem if the ref was a branch).

All in all, I think having the animated GIF displayed correctly at PyPI
is worth the cost.