calculate "skip silence" CRC32
Even though the "skip silence" variant of CRC32 is unusual and not
directly supported in zlib, one can calculate it using zlib if zero
samples are removed from input data. If input data modification is
allowed, this operation can be done in place, avoiding the additional
cost of copying a large memory area. The only caveat is that "skip
silence" CRC32 must be calculated last, because the original input
data is lost when zero samples are removed.

Adapt Python wrapper for the C extension. Print "skip silence" CRC32
in arver output and in arver-ripinfo table. Add "skip silence" CRC32
checksums to tests and verify them.

The results were also tested manually by comparing them with "W/O NULL"
CRC32 values calculated by CUETools for 51 tracks from 7 CDs.

Performance impact is noticeable, but it's definitely not as bad as my
initial attempts based on eac_crc32 code. This largely depends on the
input format: for FLAC files where the majority of processing time is
spent decoding FLAC to WAV, it's not really noticeable. It is clearly
visible for WAV files, though.

Performance testing based on running arver-ripinfo 50 times in a loop
with the same input data (~473 MiB):

        no CRCSS    with CRCSS    ratio

WAV      33.820s       56.131s     1.66
FLAC    169.672s      192.387s     1.13

This is not representative of real-life usage - in normal usage the
checksums will only be calculated once or twice.