verify source tarball checksums when building image
Verify SHA-256 checksums of tarballs after downloading. The checksums
were extracted from respective SHA256SUMS files in downloads.xiph.org.
The exception is libsndfile, where I had to download the tarball from
GitHub and calculate the checksum myself.

Checksum verification should in principle be possible in ADD command,
but it appears that I need a newer version of Docker daemon for that.