implement function for getting the number of audio frames in a file
The new function verifies the audio format first, and raises TypeError
if unsupported format is detected. The number of audio frames is only
returned if the audio data is in one of supported formats.

ARver only needs the number of audio frames to calculate the number of
CD frames (i.e. CD sectors). Other than that, it just needs to know that
the audio data it is processing is compliant with CDDA format.

Provide a Python wrapper for the C extension. The wrapper does not catch
possible exceptions to allow re-raising them as AudioFormatError in the
rip module.