include libsndfile version in ARver version string
Report version as "ARver-x.y.z (libsndfile-a.b.c)" when -v/--version
switch is given on the command line. a.b.c will either correspond to
version of libsndfile bundled in the wheel package, or to libsndfile
installed in user's OS (depending on the installation method).

This only affects version strings printed to the console by the tools.
User agent string is not changed and still contains the short version.