make sure ripped audio files match CD TOC
Add a sanity check to Rip.verify() method to report an error when ripped
files don't match the disc they're being verified against. This was not
a problem in my usage so far, but mistakes can happen and can result in
runtime crashes.

Difference in track numbers is always fatal. Mismatches in track lengths
can be ignored with the new -p/--permissive switch. This is implemented
to allow verifying tracks that only differ by some digital silence at
the end (zero value samples don't contribute to AccurateRip checksum).