protect Rip.calculate_checksums()
Having to call this method before rip information can be printed is
clumsy and error-prone. Make the method private, and call it internally
in __str__(). Make sure that checksum calculation is only done once: it
makes no sense to do so repeatably for a given set of files.

Remove explicit call of calculate_checksums() from arver_ripinfo.py.