display rip info when AccurateRip data is not available
When AccurateRip data for the disc is not found, display an error
message that's more user-friendly than the default exception string
provided by requests package. Calculate file checksums and display
rip info as a table (same as arver-ripinfo output).

Handle exceptions in AccurateRipFetcher in more detailed way. Of all
possible HTTP errors only 404 error means that the disc was not found
in the database, so print a custom message on HTTP 404 error and use the
default requests error messages for all others. Handle errors on parsing
binary database response separately.