remove exception handling from Python wrappers
Don't catch exceptions raised by the C extension in the Python wrappers,
allow all exceptions to go through. Returning zeros or None on errors
makes no sense, as it only hides edge cases that the application can't
handle correctly anyway.

Errors may occur when the file is not an audio file or it contains audio
data in unsupported format, but it doesn't happen in practice. All files
are implicitly checked in AudioFile initializer by get_nframes() call,
and any unsupported files or non-audio files are discarded.

Realistically the only exception that can occur is ValueError, when
accuraterip_checksums() receives an invalid track_number or total_tracks
argument. But in that case there's no point in handling this exception
either: this can only happen on an internal logic error (i.e. a bug). If
that happens the application should crash in a way that's visible to the
end user.