avoid sending requests to MusicBrainz in tests
get_releases_by_discid() returns release data in JSON format, so this
data can be easily stored in a file. Store information about several
discs as test data, and mock get_releases_by_discid() calls to return
test release data loaded from JSON files.

Change one of the discs used in test: its disc ID began with a period,
so its JSON file would be hidden. The new disc has the maximum number
of tracks possible in an audio CD (in addition to a pregap).

Remove assertion for checking if the returned object is not None. This
won't happen if the test is not sending requests to MusicBrainz anymore.

Avoiding network overhead makes the test execute almost instantly
instead of taking a few seconds. I'm still not sure about the idea
of this test - it seems to me that it doesn't actually test very much.