test CRC32 calculation on big endian audio data
Big endian data should now always be converted to little endian
before calculating checksums, so the checksums must be the same
regardless of audio data endianness.

It is sufficient to test this with CRC32 only - if this works then
AccurateRip checksums must be correct as well. There's no point in
duplicating tests.