set up repository mirroring
Set up GitHub Action to automatically sync changes to master branch to
ARver mirrors at GitLab and Codeberg.

It is not possible to sync both GitLab and Codeberg using one checkout,
because the action attempts to create a remote named "mirror" twice in
the same repository clone, which fails. For this reason two separate
sync jobs are required.

Checking host keys is disabled because of yesolutions/mirror-action#14.