test loading samples from a corrupted audio file
Verify that the correct exception is raised by the C extension. I'm not
sure why I didn't cover this code path before. It only makes sense to
test this path in checksums(), because nframes() does not try to load
the samples (it only reads the header).

corrupted.flac is a copy of sample.flac where some of the values were
overwritten with garbage in a hex editor, so that testing the file with
"flac -t" results in an error.