build a limited API wheel for CPython 3.7 and above
Set the options required for building a limited API wheel compatible
with CPython 3.7 and newer. This includes the three crucial parts:

- setting the Py_LIMITED_API macro to the minimum supported CPython
  version,
- adding the py_limited_api argument to Extension() call to add abi3
  tag to the wheel file name,
- adding the py_limited_api option to bdist_wheel to add the cp37 tag
  to the wheel file name.

All these options are defined in setup.py to keep them in one place,
even though the Py_LIMITED_API macro could've been defined in the C
source file.

None of this is well-documented. The following links are helpful:

https://blog.trailofbits.com/2022/11/15/python-wheels-abi-abi3audit/
https://github.com/pypa/packaging.python.org/issues/1376