adapt build script to building a single wheel
With the introduction of Python limited API we're building wheel only
with Python 3.7. We're not iterating over available Python interpreters
anymore and there's no need to remove symlinks to unsupported versions.

Change the wheel directory so that both source distribution and the
wheel end up in dist/. Remove the initial non-manylinux wheels, but
don't remove entire directory before building, only remove the wheel
files. Otherwise the source distribution would be removed as well.

Simplify iterating over wheel files. Finding and sorting is not needed
anymore, because only a single wheel file is expected.