wrap calculated checksums in a dataclass
Using a tuple of four checksums is inconvenient, so create a dedicated
Checksums type and return a Checksums object from get_checksums(). This
allows reducing the number of local variables in Rip.verify(), and
fixes a warning raised by Pylint.

Referencing the checksums as a dataclass attribute instead of directly
referencing them as local variables makes some lines longer and results
in awkward line breaks by YAPF, but this code will have to be refactored
and reformatted anyway.

Adapt unit tests.