remove setuptools workaround from unit test job
The workaround added in 5a83a70 is unnecessary since 1e56142 when the
unit test job started using an editable installation.

Editable installation is handled by pip, which has dedicated logic for
supporting legacy projects. It figures out that setuptools is a de facto
dependency and installs it first.

This is not the case when invoking Python directly to build the C
extension with the now deprecated usage of "python3 setup.py build_ext".
In that scenario nothing ensures setuptools is present, and the build
fails with Python 3.12 and newer.