remove disc.info dependency from rip
The runtime dependency is removed by moving the logic of checking if the
disc is a Mixed Mode CD to a method of DiscInfo class. This also allows
removing one local variable from verify().

The dependency on the type is removed by importing the DiscInfo class
conditionally, which satisfies MyPy. Stringifying the type in the type
hint is necessary for Pylance.

Using future annotations doesn't help here: while MyPy and Pylance seem
to be happy, Pylint complains.